requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
pyEDFlib==0.1.17
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
progressbar2